requirements.txt
python-telegram-bot
Flask
python-dotenv
psycopg2-binary